`src/func_ov006_020fce04.c` conflicted only because the two sides edited
ADJACENT declaration lines, not the same one. Both edits are kept:

  this branch  func_02012718's first parameter, void * -> int, and the
               matching `(void *)` cast dropped from the call
  main         data_ov006_0213d954 respelled from a scalar to `[]`, and the
               use respelled to `[0]`

The body had already merged to exactly that combination; only the two-line
declaration block needed resolving, and neither side touched the other's line.

`nearmiss/db.jsonl` merged through the union driver: 29 rows, no duplicate
`(module, addr)` keys, key set exactly main's union this branch's.
`config/match_attempts.jsonl` has no byte-identical duplicated lines.

+++ include/daKpFr_c.h
@@ -79,7 +79,7 @@ struct daKpFr_c : dActor_c {
     virtual void OnPendingDestroy();    /* slot 12 */
     virtual int  OnYoshiTryEat();       /* slot 18 */
 
-    static void *operator new(unsigned long size) {
+    static void *operator new(size_t size) {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
 };
