+++ include/daObjCtMecha05_c.h
@@ -31,7 +31,7 @@ struct daObjCtMecha05_c : dBgActor_c {
     int Behavior();
     int Render();
 
-    static void *operator new(unsigned long size) {
+    static void *operator new(size_t size) {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
