+++ include/daObjKinokoTag_c.h
@@ -59,7 +59,7 @@ struct daObjKinokoTag_c : dActor_c {
     /* Leaf operator new until #2570 puts the same allocator on fBase_c.
        Parameter is unsigned long (size_t on this ABI). `return new`
        relocates to `_Znwm` without this. */
-    static void *operator new(unsigned long size)
+    static void *operator new(size_t size)
     {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
