+++ include/daObjWc_Obj02_c.h
@@ -43,7 +43,7 @@ struct daObjWc_Obj02_c : dBgActor_c {
     int Behavior();
     int Render();
 
-    static void *operator new(unsigned long size)
+    static void *operator new(size_t size)
     {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
