+++ include/daObjWc_Obj05_c.h
@@ -53,7 +53,7 @@ struct daObjWc_Obj05_c : dBgActor_c {
     virtual s32   Behavior();              /* slot  6 */
     virtual s32   Render();                /* slot  9 */
 
-    static void *operator new(unsigned long size)
+    static void *operator new(size_t size)
     {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
