+++ include/daPgDfdr_c.h
@@ -59,7 +59,7 @@ struct daPgDfdr_c : dBgActor_c {
     virtual s32 Render();              /* slot  9 */
     virtual void OnPendingDestroy();   /* slot 12 -- empty in the ROM */
 
-    static void *operator new(unsigned long size) {
+    static void *operator new(size_t size) {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
